+++ cpp/include/copc-lib/io/laz_writer.hpp
@@ -33,6 +33,9 @@ class LazWriter : public BaseWriter
     {
         return std::dynamic_pointer_cast<las::LazConfigWriter>(config_);
     }
+    
+    uint64_t PointCount() { return point_count_; }
+    uint64_t ChunkCount() { return chunks_.size(); }
 };
 
 class LazFileWriter : BaseFileWriter, public LazWriter